requirements.txt
pandas
drain3
groq
pyyaml
matplotlib
numpy